// src/modules/notifications/enums/notification-type.enum.ts
export enum NotificationType {
  REPAIR_ACCEPTED = 'REPAIR_ACCEPTED',
  DIAGNOSIS_COMPLETED = 'DIAGNOSIS_COMPLETED',
  PRICE_APPROVAL_REQUIRED = 'PRICE_APPROVAL_REQUIRED',
  REPAIR_APPROVED = 'REPAIR_APPROVED',
  REPAIR_STARTED = 'REPAIR_STARTED',
  REPAIR_READY = 'REPAIR_READY',
  REPAIR_DELIVERED = 'REPAIR_DELIVERED',
  PAYMENT_RECEIVED = 'PAYMENT_RECEIVED',
  DEBT_REMINDER = 'DEBT_REMINDER',
}